fail build on missing .blend file
blender seems to always quit with exit status 0. Within the python script, we must therefore handle errors with sys.exit(1) (and appropriate logging), but in the case of the .blend file to load, the script doesn't execute before blender fails, and maven falsely assumes success. Solution: use maven-enforcer-plugin to fail even before the blender execution if the .blend file is missing.